+++ src/components/swap/CreateSwap.tsx
@@ -92,11 +92,11 @@ export const CreateSwap = () => {
     if (!connector || !inputAsset || !outputAsset) return true;
     if (!WALLET_SUPPORTED_CHAINS[connector.id]) return true;
 
-    if (
-      (connector.id === "app.phantom" || connector.id === "keplr") &&
-      outputAsset.chain === "core"
-    ) {
-      return false;
+    if (outputAsset.chain === "core") {
+      console.log(
+        WALLET_SUPPORTED_CHAINS[connector.id].includes(outputAsset.chain)
+      );
+      return WALLET_SUPPORTED_CHAINS[connector.id].includes(outputAsset.chain);
     }
 
     if (